refactor: split contract.rs into types, events, and impl modules (#225)
Move all #[contracttype] structs and service constants to src/types.rs.
Move all event structs to src/events.rs.
Reduce src/contract.rs to only the #[contractimpl] block.
Update src/lib.rs to declare the new modules and fix re-exports.
Update test files that imported types directly from crate::contract.
