+++ src/types.rs
@@ -1,0 +1,241 @@
+use soroban_sdk::{contracttype, Address, Bytes, String, Vec};
+
+// ---------------------------------------------------------------------------
+// Service constants
+// ---------------------------------------------------------------------------
+
+pub const SERVICE_DEPOSITS: u32 = 1;
+pub const SERVICE_WITHDRAWALS: u32 = 2;
+pub const SERVICE_QUOTES: u32 = 3;
+pub const SERVICE_KYC: u32 = 4;
+
+/// Typed representation of a service capability an anchor can support.
+///
+/// Each variant maps to a stable `u32` discriminant stored on-chain.
+/// Use [`ServiceType::as_u32`] to convert before passing to contract functions.
+#[derive(Clone, PartialEq)]
+pub enum ServiceType {
+    Deposits,
+    Withdrawals,
+    Quotes,
+    KYC,
+}
+
+impl ServiceType {
+    pub fn as_u32(&self) -> u32 {
+        match self {
+            ServiceType::Deposits => SERVICE_DEPOSITS,
+            ServiceType::Withdrawals => SERVICE_WITHDRAWALS,
+            ServiceType::Quotes => SERVICE_QUOTES,
+            ServiceType::KYC => SERVICE_KYC,
+        }
+    }
+}
+
+// ---------------------------------------------------------------------------
+// Core types
+// ---------------------------------------------------------------------------
+
+#[contracttype]
+#[derive(Clone)]
+pub struct Session {
+    pub session_id: u64,
+    pub initiator: Address,
+    pub created_at: u64,
+    pub nonce: u64,
+    pub operation_count: u64,
+}
+
+#[contracttype]
+#[derive(Clone)]
+pub struct Quote {
+    pub quote_id: u64,
+    pub anchor: Address,
+    pub base_asset: String,
+    pub quote_asset: String,
+    pub rate: u64,
+    pub fee_percentage: u32,
+    pub minimum_amount: u64,
+    pub maximum_amount: u64,
+    pub valid_until: u64,
+}
+
+#[contracttype]
+#[derive(Clone)]
+pub struct OperationContext {
+    pub session_id: u64,
+    pub operation_index: u64,
+    pub operation_type: String,
+    pub timestamp: u64,
+    pub status: String,
+    pub result_data: u64,
+}
+
+#[contracttype]
+#[derive(Clone)]
+pub struct AuditLog {
+    pub log_id: u64,
+    pub session_id: u64,
+    pub actor: Address,
+    pub operation: OperationContext,
+}
+
+#[contracttype]
+#[derive(Clone)]
+pub struct RequestId {
+    pub id: Bytes,
+    pub created_at: u64,
+}
+
+#[contracttype]
+#[derive(Clone)]
+pub struct Attestation {
+    pub id: u64,
+    pub issuer: Address,
+    pub subject: Address,
+    pub timestamp: u64,
+    pub payload_hash: Bytes,
+    pub signature: Bytes,
+}
+
+#[contracttype]
+#[derive(Clone)]
+pub struct TracingSpan {
+    pub request_id: RequestId,
+    pub operation: String,
+    pub actor: Address,
+    pub started_at: u64,
+    pub completed_at: u64,
+    pub status: String,
+}
+
+#[contracttype]
+#[derive(Clone)]
+pub struct AnchorServices {
+    pub anchor: Address,
+    pub services: Vec<u32>,
+}
+
+// ---------------------------------------------------------------------------
+// Routing types
+// ---------------------------------------------------------------------------
+
+#[contracttype]
+#[derive(Clone)]
+pub struct RoutingAnchorMeta {
+    pub anchor: Address,
+    pub reputation_score: u32,
+    pub average_settlement_time: u64,
+    pub liquidity_score: u32,
+    pub uptime_percentage: u32,
+    pub total_volume: u64,
+    pub is_active: bool,
+}
+
+#[contracttype]
+#[derive(Clone)]
+pub struct RoutingRequest {
+    pub base_asset: String,
+    pub quote_asset: String,
+    pub amount: u64,
+    pub operation_type: u32,
+}
+
+#[contracttype]
+#[derive(Clone)]
+pub struct RoutingOptions {
+    pub request: RoutingRequest,
+    pub strategy: Vec<soroban_sdk::Symbol>,
+    pub min_reputation: u32,
+    pub max_anchors: u32,
+    pub require_kyc: bool,
+}
+
+// ---------------------------------------------------------------------------
+// Metadata cache types
+// ---------------------------------------------------------------------------
+
+#[contracttype]
+#[derive(Clone)]
+pub struct AnchorMetadata {
+    pub anchor: Address,
+    pub reputation_score: u32,
+    pub liquidity_score: u32,
+    pub uptime_percentage: u32,
+    pub total_volume: u64,
+    pub average_settlement_time: u64,
+    pub is_active: bool,
+}
+
+#[contracttype]
+#[derive(Clone)]
+pub struct MetadataCache {
+    pub metadata: AnchorMetadata,
+    pub cached_at: u64,
+    pub ttl_seconds: u64,
+}
+
+#[contracttype]
+#[derive(Clone)]
+pub struct CapabilitiesCache {
+    pub toml_url: String,
+    pub capabilities: String,
+    pub cached_at: u64,
+    pub ttl_seconds: u64,
+}
+
+// ---------------------------------------------------------------------------
+// Anchor Info Discovery types
+// ---------------------------------------------------------------------------
+
+#[contracttype]
+#[derive(Clone)]
+pub struct AssetInfo {
+    pub code: String,
+    pub issuer: String,
+    pub deposit_enabled: bool,
+    pub withdrawal_enabled: bool,
+    pub deposit_fee_fixed: u64,
+    pub deposit_fee_percent: u32,
+    pub withdrawal_fee_fixed: u64,
+    pub withdrawal_fee_percent: u32,
+    pub deposit_min_amount: u64,
+    pub deposit_max_amount: u64,
+    pub withdrawal_min_amount: u64,
+    pub withdrawal_max_amount: u64,
+}
+
+#[contracttype]
+#[derive(Clone)]
+pub struct StellarToml {
+    pub version: String,
+    pub network_passphrase: String,
+    pub accounts: Vec<String>,
+    pub signing_key: String,
+    pub currencies: Vec<AssetInfo>,
+    pub transfer_server: String,
+    pub transfer_server_sep0024: String,
+    pub kyc_server: String,
+    pub web_auth_endpoint: String,
+}
+
+#[contracttype]
+#[derive(Clone)]
+pub struct CachedToml {
+    pub toml: StellarToml,
+    pub cached_at: u64,
+    pub ttl_seconds: u64,
+}
+
+// ---------------------------------------------------------------------------
+// Health monitoring types
+// ---------------------------------------------------------------------------
+
+#[contracttype]
+#[derive(Clone)]
+pub struct HealthStatus {
+    pub anchor: Address,
+    pub latency_ms: u64,
+    pub failure_count: u32,
+    pub availability_percent: u32,
+}
